update id increase

+++ chuyen-tau-xanh/database/migrations/2024_11_22_140357_train_routes.php
@@ -9,8 +9,8 @@ class TrainRoutes extends Migration
     public function up(): void
     {
         Schema::create('train_routes', function (Blueprint $table) {
-            $table->integer('train_id');
-            $table->integer('route_id');
+            $table->unsignedBigInteger('train_id');
+            $table->unsignedBigInteger('route_id');
             $table->integer('train_index');
             $table->timestamps();
